Rename Body component to Wrapper
