+++ src/pages/_app.tsx
@@ -1,18 +1,18 @@
 import React from "react";
 import { Container, AppProps } from "next/app";
+import Wrapper from "../components/Wrapper";
 import Header from "../components/Header";
-import Body from "../components/Body";
 import Content from "../components/Content";
 
 const MyApp = ({ Component, pageProps }: AppProps) => {
   return (
     <Container>
-      <Body>
+      <Wrapper>
         <Header />
         <Content>
           <Component {...pageProps} />
         </Content>
-      </Body>
+      </Wrapper>
     </Container>
   );
 };
